+++ src/SMAPI/PatchMode.cs
@@ -6,7 +6,7 @@ namespace StardewModdingAPI
         /// <summary>Erase the original content within the area before drawing the new content.</summary>
         Replace,
 
-        /// <summary>Draw the new content over the original content, so the original content shows through any transparent pixels.</summary>
+        /// <summary>Draw the new content over the original content, so the original content shows through any transparent or semi-transparent pixels.</summary>
         Overlay
     }
 }